Merge cumulative, for profile db change

+++ backend/Social/SocialAPI.py
@@ -9,6 +9,7 @@ import jwt
 import sys
 from datetime import datetime, timedelta
 from bson.objectid import ObjectId
+from bson.errors import *
 from flask import Blueprint
 import sys
 sys.path.append("../db")
@@ -29,39 +30,36 @@ def hello():
     return "Welcome the Raffles Hall Social server"
 
 
-#@social_api.route('/profiles', methods=['GET', 'PUT'])
 @social_api.route('/profiles', methods=['PUT'])
 @cross_origin(supports_credentials=True)
 def profiles():
     try:
-        '''
-        if request.method == 'GET':
-            data = db.Profiles.find()
+        data = request.get_json()
+        if "profilePictureURI" in data:
+            data["profilePictureUrl"] = data.pop("profilePictureURI")
+
+        if not data.get("userID"):
+            return {"status":" failed", "message": "Invalid request"}, 400
+
+        if db.User.count_documents({"userID": data["userID"]}) == 0:
+            return {"status":" failed", "message": "User not found"}, 404
+
+
+        result = db.User.update_one(
+            {"userID": data["userID"]}, {'$set': data}, upsert=True)
+
+
+        if int(result.matched_count) > 0:
             response = {
                 "status": "success",
-                "data": json.dumps(list(data), default=lambda o: str(o))
+                "message": "Profile changed"
             }
             return make_response(response, 200)
-        '''
-        if request.method == 'PUT':
-            data = request.get_json()
-            if "profilePictureURI" in data:
-                data["profilePictureUrl"] = data.pop("profilePictureURI")
-
-            result = db.Profiles.update_one(
-                {"userID": data["userID"]}, {'$set': data}, upsert=True)
-
-            if int(result.matched_count) > 0:
-                response = {
-                    "status": "success",
-                    "message": "Event changed"
-                }
-                return make_response(response, 200)
-            else:
-                response = {
-                    "status": "failed",
-                }
-                return make_response(response, 204)
+        else:
+            response = {
+                "status": "failed",
+            }
+            return make_response(response, 404)
     except Exception as e:
         print(e)
         return {"err": "An error has occured", "status": "failed"}, 500
@@ -72,15 +70,15 @@ def profiles():
 def users():
     userIdList = request.args.getlist('userID')
     try:
-        data = db.Profiles.find({"userID": {'$in': userIdList}}, {"_id": 0})
+        data = list(db.User.find({"userID": {'$in': userIdList}}, {"_id": 0, "passwordHash": 0}))
         response = {
-                 "data": list(data),
+                 "data": data,
                  "status": "success"
              }
         print(userIdList)
         if len(userIdList) == 1 and userIdList[0] == "":
             return make_response({"err": "userID is not specified", "status": "failed"}), 400 # throws error if userID is not specified in argument
-        elif db.Profiles.count_documents({"userID": {'$in': userIdList}}) == 0:
+        elif len(data) == 0:
             return make_response({"err": "User does not exist", "status": "failed"}), 404 # throws error if all userID entries in argument do not exist 
         else:
             return make_response(response, 200)
@@ -92,18 +90,22 @@ def users():
 @cross_origin(supports_credentials=True)
 def getUserProfile(userID):
     try:
-        data = db.Profiles.find({"userID": userID}, {"_id": 0})
+        if (not userID):
+            return {"err": "Invalid request"}, 400
+        data = list(db.User.find({"userID": userID}, {"_id": 0, "passwordHash": 0}))
         response = {
-            "data": list(data),
+            "data": data,
             "status": "success"
         }
+
+        if len(data) == 0:
+            return make_response({"err": "User does not exist", "status":"failed"}), 404
+        else:
+            return make_response(response, 200)
     except Exception as e:
         print(e)
         return {"err": "An error has occured", "status": "failed"}, 500
-    if db.Profiles.count_documents({"userID": userID}) == 0:
-        return make_response({"err": "User does not exist", "status":"failed"}), 404
-    else:
-        return make_response(response, 200)
+    
         
 
 @social_api.route("/user", methods=['PUT', 'POST'])
@@ -127,7 +129,7 @@ def user():
 
             if int(result.matched_count) > 0:
                 response = {
-                    "message": "Event changed",
+                    "message": "Profile changed",
                     "status": "success"
                 }
                 return make_response(response, 200)
@@ -135,23 +137,26 @@ def user():
                 response = {
                     "status": "failed"
                 }
-                return make_response(response, 204)
+                return make_response(response, 400)
 
         elif request.method == 'POST':
             data = request.get_json()
-            userID = str(data.get('userID'))
-            passwordHash = str(data.get('passwordHash'))
-            email = str(data.get('email'))
-            position = []  # default to be empty, will be added manually from BE
+            userID = data.get('userID')
+            passwordHash = data.get('passwordHash')
+            email = data.get('email')
+
+            if (userID is None or passwordHash is None or email is None):
+                return {"status": "failed", "message": "Invalid request"}, 400
+
+            if db.Users.count_documents({"userID": userID}) == 0:
+                return {"status": "failed", "message": "Account already exist"}, 409
 
             body = {
                 "userID": userID,
                 "passwordHash": passwordHash,
                 "email": email,
-                "position": position
             }
-            receipt = db.User.insert_one(body)
-            body["_id"] = str(receipt.inserted_id)
+            db.User.insert_one(body)
 
             return make_response({"message": body, "status": "success"}, 200)
 
@@ -160,7 +165,7 @@ def user():
         return {"err": "An error has occured", "status": "failed"}, 500
 
 
-@social_api.route("/user/<userID>")
+@social_api.route("/user/<userID>", methods=["GET"])
 @cross_origin(supports_credentials=True)
 def getUserDetails(userID):
     try:
@@ -170,7 +175,7 @@ def getUserDetails(userID):
             }
             },
             {'$lookup': {
-                'from': 'Profiles',
+                'from': 'User',
                         'localField': 'userID',
                         'foreignField': 'userID',
                         'as': 'profile'
@@ -179,7 +184,7 @@ def getUserDetails(userID):
             {
                 '$replaceRoot': {'newRoot': {'$mergeObjects': [{'$arrayElemAt': ["$profile", 0]}, "$$ROOT"]}}
             },
-            {'$project': {'profile': 0}},
+            {'$project': {'profile': 0, "_id": 0}},
             {'$lookup': {
                 'from': 'CCA',
                         'localField': 'position',
@@ -187,17 +192,14 @@ def getUserDetails(userID):
                         'as': 'positions'
             }
             },
-            {'$project': {'positions.category': 0, 'positions._id': 0, 'position': 0}}
+            {'$project': {'positions.category': 0, 'positions._id': 0, 'position': 0, 'passwordHash': 0}}
         ]
 
-        data = db.User.aggregate(pipeline)
-        response = None
-        for item in data:
-            response = item
+        data = list(db.User.aggregate(pipeline))
 
         return make_response(
             {
-                "data": json.dumps(response, default=lambda o: str(o)),
+                "data": data,
                 "status": "success"
             }, 200)
 
@@ -205,18 +207,11 @@ def getUserDetails(userID):
         print(e)
         return {"err": "An error has occured", "status": "failed"}, 500
 
-    return make_response(
-        {
-            "data": json.dumps(response, default=lambda o: str(o)),
-            "status": "success"
-        },
-        200)
-
 
 def userIDtoName(userID):
     # TODO use mongoDB lookup instead of this disgusting code
     # helper function
-    profile = db.Profiles.find_one({"userID": userID})
+    profile = db.User.find_one({"userID": userID}, {'passwordHash': 0, 'displayName': 1})
     name = profile.get('displayName') if profile else None
     return name
 
@@ -233,8 +228,8 @@ def posts():
                 if postID:
                     # TODO use mongoDB lookup to join the data instead
                     data = db.Posts.find_one({"_id": ObjectId(postID)})
-                    name = db.Profiles.find_one(
-                        {"userID": str(data.get("userID"))}).get('displayName')
+                    name = db.User.find_one(
+                        {"userID": str(data.get("userID"))}, {'passwordHash': 0}).get('displayName')
 
                     if data != None:
                         data = renamePost(data)
@@ -283,7 +278,7 @@ def posts():
                 #     {'limit': 5},
                 #     {
                 #         '$lookup': {
-                #             'from': 'Profiles',
+                #             'from': 'User',
                 #             'localField': 'userID',
                 #             'foreignField': 'userID',
                 #             'as': 'profile'
@@ -298,7 +293,7 @@ def posts():
                 #             'name': '$profile.displayName'
                 #         }
                 #     },
-                #     {'$project': {'profile': 0}}
+                #     {'$project': {'profile': 0} }
                 # ]
 
                 # data = db.Posts.aggregate(pipeline)
@@ -306,7 +301,7 @@ def posts():
                 response = []
 
                 userIDList = [x["userID"] for x in data]
-                profiles = list(db.Profiles.find({"userID": {"$in": userIDList}}))
+                profiles = list(db.User.find({"userID": {"$in": userIDList}}, {'passwordHash': 0}))
                 profileDict = {}
                 for x in profiles:
                     profileDict[x["userID"]] = x
@@ -329,12 +324,23 @@ def posts():
                     }, 200)
 
         elif request.method == 'DELETE':
-            postID = request.args.get('postID')
-            db.Posts.delete_one({"_id": ObjectId(postID)})
-            response = {
-                "status": "success"
-            }
-            return make_response(response, 200)
+            try:
+                postID = request.args.get('postID')
+                if db.Posts.count_documents({"_id": ObjectId(postID)}) == 0:
+                    response = {
+                        "status": "failed",
+                        "message": "The post you are trying to delete does not exist."
+                    }
+                    return make_response(response, 404)
+                else:
+                    db.Posts.delete_one({"_id": ObjectId(postID)})
+                    response = {
+                        "status": "success",
+                        "message": "Post deleted successfully"
+                    }
+                    return make_response(response, 200)
+            except InvalidId as e:
+                return make_response({"status": "failed", "message": "Invalid postID"}), 400
 
         elif request.method == 'POST':
             data = request.get_json()
@@ -346,6 +352,9 @@ def posts():
             postPics = data.get('postPics') if data.get('postPics') else []
             isOfficial = bool(data.get('isOfficial'))
             tags = data.get('tags')
+
+            if (not userID or userID == ""):
+                return {"status":" failed", "message": "Invalid request"}, 400
 
             body = {
                 "userID": userID,
@@ -366,10 +375,16 @@ def posts():
         elif request.method == 'PUT':
             data = request.get_json()
             postID = data.get('postID')
-            oldPost = db.Posts.find_one({"_id": ObjectId(postID)})
+            if (data.get('postID') is None):
+                return {"status":" failed", "message": "Invalid request"}, 400
 
-            if oldPost == None:
-                return make_response("data non existent", 404)
+            try:
+                ObjectId(postID)
+            except (InvalidId):
+                return {"status":" failed", "message": "Invalid request"}, 400
+
+            if db.Posts.count_documents({"_id": ObjectId(postID)}) == 0:
+                return make_response({"status":" failed", "message": "Post does not exist"}, 404)
 
             userID = str(data.get('userID')) if data.get(
                 'userID') else oldPost.get('userID')
@@ -396,9 +411,9 @@ def posts():
             result = db.Posts.update_one(
                 {"_id": ObjectId(postID)}, {'$set': body})
             if int(result.matched_count) > 0:
-                return make_response({'message': "Event changed"}, 200)
+                return make_response({'message': "Post changed", "status": "success"}, 200)
             else:
-                return Response(status=204)
+                return Response(status=200)
     except Exception as e:
         print(e)
         return {"err": "An error has occured", "status": "failed"}, 500
@@ -496,7 +511,7 @@ def getOfficialPosts():
         for item in data:
             item['name'] = userIDtoName(item.get('userID'))
             ccaID = int(item.get('ccaID'))
-            profile = db.Profiles.find_one({'userID': item.get('userID')})
+            profile = db.User.find_one({'userID': item.get('userID')}, {'passwordHash': 0})
             item['profilePictureURI'] = profile.get(
                 'profilePictureUrl') if profile != None else None
             item['ccaName'] = db.CCA.find_one({'ccaID': ccaID}).get(
@@ -529,7 +544,7 @@ def getAllFriends(userID):
         result = []
 
         for friendID in friends:
-            entry = db.Profiles.find_one({"userID": friendID})
+            entry = db.User.find_one({"userID": friendID}, {'passwordHash': 0})
             if entry != None:
                 result.append(entry)
 
@@ -543,3 +558,65 @@ def getAllFriends(userID):
     except Exception as e:
         print(e)
         return {"err": "An error has occured", "status": "failed"}, 500
+
+@social_api.route('/cca/<int:ccaID>', methods=["GET"])
+@cross_origin()
+def getCCADetails(ccaID):
+    try:
+        data = list(db.CCA.find({"ccaID": ccaID}, {'_id': 0}))
+        response = {"status": "success", "data": data}
+    except Exception as e:
+        print(e)
+        return {"err": "An error has occured", "status": "failed"}, 500
+    return make_response(response)
+
+
+@social_api.route("/user_CCA/<string:userID>", methods=['GET'])
+@cross_origin()
+def getUserCCAs(userID):
+    try:
+        CCAofUserID = db.UserCCA.find({"userID": userID})
+        entries = [w["ccaID"] for w in CCAofUserID]
+        data = list(db.CCA.find({"ccaID": {"$in": entries}}, {"_id": 0}))
+        response = {"status": "success", "data": data}
+    except Exception as e:
+        print(e)
+        return {"err": "An error has occured", "status": "failed"}, 500
+    return make_response(response, 200)
+
+@social_api.route("/user_CCA", methods=['POST', 'DELETE'])
+@cross_origin()
+def editUserCCA():
+    try:
+        data = request.get_json()
+        ccaID = data.get('ccaID')
+        userID = data.get('userID')
+
+        if request.method == "POST":
+            db.UserCCA.delete_many({"userID": userID})
+            if len(ccaID) > 0:
+                documents = [{"userID": userID, "ccaID": cca} for cca in ccaID]
+                upserts = [pymongo.UpdateOne(
+                    doc, {'$setOnInsert': doc}, upsert=True) for doc in documents]
+                db.UserCCA.bulk_write(upserts)
+
+        elif request.method == "DELETE":
+            db.UserCCA.delete_many(body)
+
+    except Exception as e:
+        print(e)
+        return {"err": "An error has occured", "status": "failed"}, 500
+    return {"status": "success"}, 200
+
+
+@social_api.route("/user_CCA/", methods=["GET"])
+@cross_origin()
+def getCCAMembersName():
+    try:
+        ccaName = str(request.args.get('ccaName'))
+        response = db.UserCCA.find({"ccaName": ccaName})
+        response = {"status": "success", "data": list(response)}
+    except Exception as e:
+        print(e)
+        return {"err": "An error has occured", "status": "failed"}, 500
+    return make_response(response)
